+++ Mods/Project Eternity Sorcerer Street/Movement Algorithm.cs
@@ -23,7 +23,7 @@ namespace ProjectEternity.GameScreens.SorcererStreetScreen
 
             foreach (MovementAlgorithmTile ActiveDestination in ListLayerPossibility)
             {
-                MovementAlgorithmTile ActiveTile = GetTile(StartingNode.Position.X + OffsetX, StartingNode.Position.X + OffsetY, ActiveDestination.LayerIndex);
+                MovementAlgorithmTile ActiveTile = GetTile((int)(StartingNode.Position.X + OffsetX), (int)(StartingNode.Position.X + OffsetY), ActiveDestination.LayerIndex);
                 //Wall
                 if (ActiveTile == null || ActiveTile.MVEnterCost == -1 || ActiveTile.MovementCost == -1
                     || ActiveTile.TerrainTypeIndex == UnitStats.TerrainWallIndex || ActiveTile.TerrainTypeIndex == UnitStats.TerrainVoidIndex)
@@ -49,14 +49,14 @@ namespace ProjectEternity.GameScreens.SorcererStreetScreen
             return TerrainToGo.MVMoveCost;
         }
 
-        public override MovementAlgorithmTile GetTile(float PosX, float PosY, int LayerIndex)
+        public override MovementAlgorithmTile GetTile(int PosX, int PosY, int LayerIndex)
         {
             if (PosX < 0 || PosY < 0 || PosX >= Map.MapSize.X || PosY >= Map.MapSize.Y)
             {
                 return null;
             }
 
-            return Map.ListLayer[LayerIndex].ArrayTerrain[(int)PosX, (int)PosY];
+            return Map.ListLayer[LayerIndex].ArrayTerrain[PosX, PosY];
         }
 
         public override bool IsBlocked(MovementAlgorithmTile CurrentNode)
